Wrote categories db function and added global styles file
